LayoutTests/inspector/debugger/breakpoints/resolved-dump-all-inline-script-pause-locations.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="../resources/log-pause-location.js"></script>
<script src="resources/dump.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("Debugger.resolvedBreakpoint.dumpAllLocations");

    window.addDumpAllPauseLocationsTestCase(suite, {
        name: "Debugger.resolvedBreakpoint.dumpAllLocations.InlineScript",
        resourceRegex: /dump-inline-scripts\.html$/,
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Checking all resolved breakpoint locations in a script.</p>
<iframe src="resources/dump-inline-scripts.html"></iframe>
</body>
</html>
